Refactored components folder
